continuing to implement satrec and associated logical flow
